Remove second style of color_Bar + PQA comments + Update notebook
